+++ tomcat/src/main/java/nextstep/jwp/controller/ErrorController.java
@@ -1,31 +1,31 @@
-package nextstep.jwp.handler;
+package nextstep.jwp.controller;
 
 import org.apache.coyote.http11.enums.HttpStatusCode;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.HttpResponse;
 
-public class FileController extends AbstractController {
+public class ErrorController extends AbstractController {
 
-    private static final Controller INSTANCE = new FileController();
+    private static final ErrorController INSTANCE = new ErrorController();
 
     public static Controller getInstance() {
         return INSTANCE;
     }
 
-    private FileController() {
+    private ErrorController() {
     }
 
     @Override
     protected HttpResponse doGet(final HttpRequest httpRequest) {
-        return generateResponse(httpRequest);
+        return generateResponse();
     }
 
     @Override
     protected HttpResponse doPost(final HttpRequest httpRequest) {
-        return generateResponse(httpRequest);
+        return generateResponse();
     }
 
-    private HttpResponse generateResponse(final HttpRequest httpRequest) {
-        return HttpResponse.of(HttpStatusCode.OK, httpRequest.getUrl());
+    private HttpResponse generateResponse() {
+        return HttpResponse.of(HttpStatusCode.NOT_FOUND, "/404.html");
     }
 }
