+++ tomcat/src/main/java/nextstep/jwp/controller/LoginController.java
@@ -1,4 +1,4 @@
-package nextstep.jwp.handler;
+package nextstep.jwp.controller;
 
 import java.util.Optional;
 import nextstep.Application;
