+++ tomcat/src/test/java/nextstep/jwp/controller/RegisterControllerTest.java
@@ -1,4 +1,4 @@
-package nextstep.jwp.handler;
+package nextstep.jwp.controller;
 
 import static org.assertj.core.api.Assertions.assertThat;
